pyproject.toml
[tool.poetry]
name = "imagebaseline"
version = "0.1.0"
description = ""
authors = ["Masataka Imamura <masataka.i.navi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.1"
pandas = "^1.1.1"
matplotlib = "^3.3.1"
seaborn = "^0.10.1"
jupyter = "^1.0.0"
scikit-learn = "^0.23.2"
pandas-profiling = "^2.8.0"
opencv-python = "^4.4.0"
albumentations = "^0.4.6"
comet_ml = "^3.2.0"
hydra-core = "^0.11.3"
category_encoders = "^2.2.2"
torch = "^1.6.0"
torchvision = "^0.7.0"
pytorch_lightning = "^0.9.0"
efficientnet-pytorch = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
